# apps/web/api/intents.ts
import type { IntentDetails, IntentStatus } from "@agent-intents/shared";
import { isSupportedChain } from "@agent-intents/shared";
/**
 * Intents endpoint
 *
 * GET  /api/intents?userId=...&status=...&limit=...&cursor=...  – List intents for a user
 * POST /api/intents                                    – Create a new intent
 *
 * POST requires AgentAuth header signed by a registered agent key.
 * The trustchain_id and member_id are derived from the verified signature.
 */
import type { VercelRequest, VercelResponse } from "@vercel/node";
import { v4 as uuidv4 } from "uuid";
import { isAddress } from "viem";
import { verifyAgentAuth } from "./_lib/agentAuth.js";
import { requireSession } from "./_lib/auth.js";
import { withDbRlsContext } from "./_lib/db.js";
import {
	authError,
	getQueryNumber,
	getQueryParam,
	jsonError,
	jsonSuccess,
	methodRouter,
	parseBodyWithSchema,
} from "./_lib/http.js";
import { createIntent, getIntentsByUser, supersedePendingIntents } from "./_lib/intentsRepo.js";
import { logger } from "./_lib/logger.js";
import { verifyAndEnrichPolymarketIntent } from "./_lib/oracleCre.js";
import { createIntentRequestSchema } from "./_lib/validation.js";

/** Maximum intents per agent per minute */
const RATE_LIMIT_PER_MINUTE = 10;

const VALID_STATUSES: IntentStatus[] = [
	"pending",
	"approved",
	"rejected",
	"broadcasting",
	"authorized",
	"confirmed",
	"failed",
	"expired",
];

export default methodRouter({
	GET: async (req: VercelRequest, res: VercelResponse) => {
		// Require authenticated session
		let session: { sessionId: string; walletAddress: string };
		try {
			session = await requireSession(req);
		} catch {
			authError(req, res, "Authentication required", 401);
			return;
		}

		const userId = getQueryParam(req, "userId");

		if (!userId) {
			jsonError(res, "Missing required query parameter: userId", 400);
			return;
		}

		// Enforce ownership: users can only list their own intents
		if (userId.toLowerCase() !== session.walletAddress) {
			authError(req, res, "You can only list your own intents", 403, session.walletAddress);
			return;
		}

		const statusParam = getQueryParam(req, "status");
		const status =
			statusParam && VALID_STATUSES.includes(statusParam as IntentStatus)
				? (statusParam as IntentStatus)
				: undefined;

		const limit = getQueryNumber(req, "limit", 50, 1, 100);
		const cursor = getQueryParam(req, "cursor");

		const page = await withDbRlsContext({ currentUser: session.walletAddress }, async (client) =>
			getIntentsByUser({ userId, status, limit, cursor }, client.sql),
		);
		jsonSuccess(res, page);
	},

	POST: async (req: VercelRequest, res: VercelResponse) => {
		const authHeader = req.headers.authorization;
		if (!authHeader?.startsWith("AgentAuth ")) {
			authError(req, res, "Missing or invalid Authorization header", 401);
			return;
		}

		let userId: string;
		let trustChainId: string | undefined;
		let createdByMemberId: string | undefined;
		try {
			const { member } = await verifyAgentAuth(req);
			userId = member.trustchainId;
			trustChainId = member.trustchainId;
			createdByMemberId = member.id;
		} catch (err) {
			authError(req, res, "Authentication failed", 401);
			return;
		}

		const body = parseBodyWithSchema(req, res, createIntentRequestSchema);
		if (body === null) return;

		// Semantic validation — branch by intent type
		if (!isSupportedChain(body.details.chainId)) {
			jsonError(res, "Unsupported chain ID", 400);
			return;
		}

		if (body.details.type === "polymarket_trade") {
			if (
				typeof body.details.amount !== "string" ||
				body.details.amount.length === 0 ||
				Number.isNaN(Number(body.details.amount))
			) {
				jsonError(res, "Invalid amount", 400);
				return;
			}
			try {
				const enriched = await verifyAndEnrichPolymarketIntent(body.details);
				(body as { details: typeof enriched }).details = enriched;
			} catch (err) {
				const message = err instanceof Error ? err.message : "Polymarket enrichment failed";
				jsonError(res, message, 400);
				return;
			}
		} else {
			if (!isAddress(body.details.recipient)) {
				jsonError(res, "Invalid recipient address", 400);
				return;
			}
			if (
				typeof body.details.amount !== "string" ||
				body.details.amount.length === 0 ||
				Number.isNaN(Number(body.details.amount))
			) {
				jsonError(res, "Invalid amount", 400);
				return;
			}
		}

		// Rate limiting: max N intents per agent per minute (fail closed)
		try {
			const rateLimitExceeded = await withDbRlsContext({ currentUser: userId }, async (client) => {
				const oneMinuteAgo = new Date(Date.now() - 60_000).toISOString();
				const countResult = await client.sql`
					SELECT COUNT(*)::int AS cnt
					FROM intents
					WHERE agent_id = ${body.agentId}
						AND created_at > ${oneMinuteAgo}
				`;
				const recentCount = (countResult.rows[0] as { cnt: number })?.cnt ?? 0;
				if (recentCount >= RATE_LIMIT_PER_MINUTE) {
					jsonError(
						res,
						`Rate limit exceeded: agent \"${body.agentId}\" has created ${recentCount} intents in the last minute (max ${RATE_LIMIT_PER_MINUTE})`,
						429,
					);
					return true;
				}
				return false;
			});
			if (rateLimitExceeded) return;
		} catch (err) {
			logger.error({ err }, "rate-limit check failed");
			jsonError(res, "Service temporarily unavailable", 503);
			return;
		}

		// Auto-expire any existing pending intents from the same agent so the
		// queue doesn't accumulate duplicates when an agent recreates intents.
		try {
			const superseded = await withDbRlsContext({ currentUser: userId }, async (client) =>
				supersedePendingIntents(body.agentId, userId, client),
			);
			if (superseded > 0) {
				logger.info(
					{ agentId: body.agentId, count: superseded },
					"Superseded pending intents from same agent",
				);
			}
		} catch (err) {
			logger.warn({ err, agentId: body.agentId }, "Failed to supersede pending intents");
		}

		// Generate ID
		const id = `int_${Date.now()}_${uuidv4().slice(0, 8)}`;

		// Calculate expiration
		const expiresAt = body.expiresInMinutes
			? new Date(Date.now() + body.expiresInMinutes * 60 * 1000).toISOString()
			: new Date(Date.now() + 24 * 60 * 60 * 1000).toISOString(); // Default 24h

		const intent = await withDbRlsContext({ currentUser: userId }, async (client) =>
			createIntent(
				{
					id,
					userId,
					agentId: body.agentId,
					agentName: body.agentName ?? body.agentId,
					details: body.details as unknown as IntentDetails,
					urgency: body.urgency ?? "normal",
					expiresAt,
					trustChainId,
					createdByMemberId,
				},
				client.sql,
			),
		);

		// Build payment URL from request headers
		const proto = req.headers["x-forwarded-proto"] ?? "https";
		const host = req.headers["x-forwarded-host"] ?? req.headers.host ?? "localhost:3000";
		const paymentUrl = `${proto}://${host}/pay/${intent.id}`;

		const logFields: Record<string, unknown> = {
			intentId: intent.id,
			agentName: intent.agentName,
			type: intent.details.type,
			amount: intent.details.amount,
		};
		if (intent.details.type === "transfer") {
			logFields.token = intent.details.token;
			logFields.recipient = intent.details.recipient;
		} else if (intent.details.type === "polymarket_trade") {
			logFields.conditionId = intent.details.conditionId;
			logFields.outcome = intent.details.outcome;
		}
		logger.info(logFields, "Intent created");

		jsonSuccess(res, { intent, paymentUrl }, 201);
	},
});